@@ -1,21 +1,28 @@
 import React from "react";
-import { Switch, Route, BrowserRouter } from "react-router-dom";
-import { StylesProvider } from "@material-ui/core/styles";
+import { Routes, Route, BrowserRouter } from "react-router-dom";
+import { StylesProvider } from "@mui/styles";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+
+const theme = createTheme();
 
 import Landing from "./components/Landing";
 import Pricing from "./components/Pricing";
 
-export default () => {
+const App = () => {
   return (
     <div>
-      <StylesProvider>
-        <BrowserRouter>
-          <Switch>
-            <Route exact path="/pricing" component={Pricing} />
-            <Route path="/" component={Landing} />
-          </Switch>
-        </BrowserRouter>
-      </StylesProvider>
+      <ThemeProvider theme={theme}>
+        <StylesProvider>
+          <BrowserRouter>
+            <Routes>
+              <Route exact path="/pricing" element={<Pricing />} />
+              <Route path="/" element={<Landing />} />
+            </Routes>
+          </BrowserRouter>
+        </StylesProvider>
+      </ThemeProvider>
     </div>
   );
 };
+
+export default App;
